Revert "Be explicit about libomp-devel denpending on libomp"
This partially reverts commit f5327ee038bb53610ce05d6e6002ea52cd37b583
because it didn't fix rhbz#2187642.
We found out the issue might be related to releng changes.
We need a new release with a new NVR in order for Bodhi and pungi to
release the i686 libomp package in the x86_64 repository.